[svn r20015] get num might return long
--HG--
branch : sod

+++ src/edu/sc/seis/sod/hibernate/SodDB.java
@@ -222,7 +222,7 @@ public class SodDB extends AbstractHibernateDB {
         query.setMaxResults(1);
         List result = query.list();
         if(result.size() > 0) {
-            return (Integer)result.get(0);
+            return ((Long)result.get(0)).intValue();
         }
         return 0;
     }
